# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agro API",
        "description": "Documentación de la API de autenticación con Swagger en Laravel 11",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/about-us": {
            "get": {
                "tags": [
                    "AboutUs"
                ],
                "summary": "Obtener información de About Us",
                "operationId": "75b59cf230787273ea2ec9a03f704023",
                "responses": {
                    "200": {
                        "description": "Información obtenida correctamente"
                    },
                    "404": {
                        "description": "Registro no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/about-us/{idAboutUs}": {
            "put": {
                "tags": [
                    "AboutUs"
                ],
                "summary": "Actualizar información de About Us",
                "operationId": "d909ed44da73d3b59ac9b4aec3995617",
                "parameters": [
                    {
                        "name": "idAboutUs",
                        "in": "path",
                        "description": "ID del registro de About Us",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mission": {
                                        "type": "string",
                                        "example": "Ser la mejor empresa del sector"
                                    },
                                    "vision": {
                                        "type": "string",
                                        "example": "Innovar y expandir nuestras soluciones"
                                    },
                                    "name_yt": {
                                        "type": "string",
                                        "example": "Nuestro canal oficial"
                                    },
                                    "url_yt": {
                                        "type": "string",
                                        "example": "https://www.youtube.com/channel/ejemplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Información actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Información actualizada correctamente"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No se enviaron datos válidos para actualizar"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/about-us/image": {
            "post": {
                "tags": [
                    "AboutUs"
                ],
                "summary": "Crea o Actualiza una imagen de About Us",
                "operationId": "e8c00b517e46f4e1a7ab6ee397a8c2ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "Imagen en base64",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Imagen actualizada con éxito"
                    },
                    "500": {
                        "description": "Error al actualizar la imagen"
                    }
                }
            }
        },
        "/api/about-us/add-value/{idValue}": {
            "put": {
                "tags": [
                    "AboutUs"
                ],
                "summary": "Actualizar un valor existente en About Us",
                "operationId": "64c6e9aa409e6c757a5aea2a579c2a49",
                "parameters": [
                    {
                        "name": "idValue",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "oldAboutValue": {
                                        "type": "string"
                                    },
                                    "newAboutValue": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Valor actualizado correctamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "AboutUs"
                ],
                "summary": "Agregar un nuevo valor a About Us",
                "operationId": "7bf09206ac09cfa719937aabb227a3dd",
                "parameters": [
                    {
                        "name": "idValue",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "about_values": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Valor agregado con éxito"
                    }
                }
            }
        },
        "/api/about-us/delete-value/{idValue}": {
            "delete": {
                "tags": [
                    "AboutUs"
                ],
                "summary": "Eliminar un valor de About Us",
                "operationId": "22e35c35a51c2c775a7a66a6fc1835a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "aboutValue": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Valor eliminado correctamente"
                    },
                    "404": {
                        "description": "Registro no encontrado"
                    },
                    "422": {
                        "description": "El valor no existe en la lista"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}